[chore #163038441] Move test folder into src
