+++ src/data/meetups.json
@@ -1,1 +1,1 @@
-[{"id":1,"title":"New title Again","location":"Lagos, Nigeria","images":[],"createdOn":"2019-01-05T17:01:34.394Z","happeningOn":"2019-03-05T17:01:34.351Z","tags":[]},{"id":2,"title":"New title Again","location":"Lagos, Nigeria","images":[],"createdOn":"2019-01-05T17:01:34.394Z","happeningOn":"2019-03-05T17:01:34.351Z","tags":[]},{"id":3,"title":"New title Again","location":"Lagos, Nigeria","images":[],"createdOn":"2019-01-05T17:01:34.394Z","happeningOn":"2019-03-05T17:01:34.351Z","tags":[]}]
+[{"id":2,"title":"New title Again","location":"Lagos, Nigeria","images":[],"createdOn":"2019-01-05T17:01:34.394Z","happeningOn":"2019-03-05T17:01:34.351Z","tags":[]},{"id":3,"title":"New title Again","location":"Lagos, Nigeria","images":[],"createdOn":"2019-01-05T17:01:34.394Z","happeningOn":"2019-03-05T17:01:34.351Z","tags":[]},{"id":4,"title":"New title Again","location":"Lagos, Nigeria","images":[],"createdOn":"2019-01-08T07:34:02.794Z","happeningOn":"2019-03-08T07:34:02.661Z","tags":[]}]
